test(completion): unit-test pure completion logic in vscode-free modules
Extract the pure logic of the completion providers into modules that do
not import `vscode` at runtime, so it can be covered by the fast Node
`test:unit` harness (properly instrumented by c8) instead of the slow
extension-host suite:

- antoraResourceId.ts: buildResourceIds + findAntoraResourceMacroPrefix
- antoraResourceMacro.ts: matchAntoraResourceMacroAt (returns offsets;
  findAntoraResourceMacroAt now wraps it to build the vscode.Range)
- xrefIdExtractor.ts: anchor id extractors (pure, synchronous)

Add unit tests for each module and drop the now-duplicated pure-function
describe blocks from the extension-host suites.
