+++ src/features/antora/antoraResourceDefinitionProvider.ts
@@ -1,18 +1,6 @@
 import * as vscode from 'vscode'
 import { getAntoraDocumentContext } from './antoraDocument.js'
-
-// Matches the `image:`/`image::`, `xref:` and `include::` macros and captures
-// their target, e.g. `image::2.0@cli:commands:output.png[]`.
-const MACRO_RX = /(image|xref|include)(::?)([^\s[\]]+)\[/g
-
-// An Antora resource id always carries a family/component/module/version marker.
-const RESOURCE_ID_DETECTOR_RX = /[$:@]/
-
-const DEFAULT_FAMILY_BY_MACRO: { [macro: string]: string } = {
-  image: 'image',
-  xref: 'page',
-  include: 'page',
-}
+import { matchAntoraResourceMacroAt } from './antoraResourceMacro.js'
 
 export interface AntoraResourceMacro {
   /** The resource id to resolve, without any `#fragment`. */
@@ -32,36 +20,18 @@ export function findAntoraResourceMacroAt(
   lineNumber: number,
   character: number,
 ): AntoraResourceMacro | undefined {
-  for (const match of lineText.matchAll(MACRO_RX)) {
-    const macro = match[1]
-    const target = match[3]
-    const targetStart = match.index + match[1].length + match[2].length
-    const targetEnd = targetStart + target.length
-    if (character < targetStart || character > targetEnd) {
-      continue
-    }
-    // Drop the fragment (e.g. `xref:page.adoc#anchor[]`) before resolution.
-    const fragmentIndex = target.indexOf('#')
-    const id = fragmentIndex === -1 ? target : target.slice(0, fragmentIndex)
-    if (id.length === 0) {
-      continue
-    }
-    // For includes, only resource ids go through the content catalog; plain
-    // relative paths are resolved by the include processor at render time.
-    if (macro === 'include' && !RESOURCE_ID_DETECTOR_RX.test(id)) {
-      continue
-    }
-    const idEnd = fragmentIndex === -1 ? targetEnd : targetStart + fragmentIndex
-    return {
-      id,
-      family: DEFAULT_FAMILY_BY_MACRO[macro],
-      range: new vscode.Range(
-        new vscode.Position(lineNumber, targetStart),
-        new vscode.Position(lineNumber, idEnd),
-      ),
-    }
+  const match = matchAntoraResourceMacroAt(lineText, character)
+  if (match === undefined) {
+    return undefined
   }
-  return undefined
+  return {
+    id: match.id,
+    family: match.family,
+    range: new vscode.Range(
+      new vscode.Position(lineNumber, match.idStart),
+      new vscode.Position(lineNumber, match.idEnd),
+    ),
+  }
 }
 
 function abspathToUri(abspath: string): vscode.Uri {
